pyproject.toml
[project]
name = "anti-fraud-kuma"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kagglehub (>=0.4.2,<0.5.0)",
    "pandas (>=2.2)",
    "numpy (>=2.1)",
    "kumoai (>=2.14.1,<3.0.0)",
    "openai (>=1.0.0)",
    "python-dotenv (>=1.0.0)",
    "fastapi (>=0.115)",
    "uvicorn (>=0.34)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
